Add files via upload
This version is a re-write of the logic to play the game, with the code for the functions copied from the last added version on 04/10.  This version needs the addition of comments and possibly some cleaning up.  Comments had been made on the version posted by Jim on 04/10.  These can be copied, or re-written if incorrect